requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
gunicorn
django-cors-headers==4.5.0
drf-spectacular==0.27.2
